+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `List` RPC in the user registry and related messages.
 - User management for admins in the Console.
 - User list command in the CLI.
+- Support for Microchip ATECC608A-TNGLORA-C manifest files in device template conversion.
 
 ### Changed